requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
beautifulsoup4==4.10.0
boto3==1.21.32
botocore==1.24.32
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
msgpack==1.0.3
mypy==0.940
mypy-extensions==0.4.3
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
types-beautifulsoup4==4.10.18
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing_extensions==4.1.1
urllib3==1.26.9
webencodings==0.5.1
zope.interface==5.4.0

